refactor employment
